IE to IC renames for API implementation but no changes to proposed API yet (#184605)
